checks into hooks
